[Refactor]Extrace common class
