+++ src/main/java/day6/IntStreamPractice.java
@@ -4,7 +4,8 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import java.util.stream.Stream;
+
+import common.Employee;
 
 public class IntStreamPractice {
     public static void main(String[] args) {
