+++ src/main/java/streamapi/basic/StreamPractice.java
@@ -3,6 +3,8 @@ package streamapi.basic;
 import java.util.ArrayList;
 import java.util.List;
 
+import common.Person;
+
 /**
  * Java SE8 StreamAPIの練習
  * 今日の課題は： no for-loop
